+++ readium/navigator/src/main/java/org/readium/r2/navigator/epub/EpubNavigatorViewModel.kt
@@ -85,14 +85,17 @@ internal class EpubNavigatorViewModel(
 
     private val _settings: MutableStateFlow<EpubSettings> =
         MutableStateFlow(settingsPolicy.settings(initialPreferences))
-
     val settings: StateFlow<EpubSettings> = _settings.asStateFlow()
 
     val overflow: StateFlow<OverflowableNavigator.Overflow> = _settings
         .mapStateIn(viewModelScope) { settings ->
             SimpleOverflow(
                 readingProgression = settings.readingProgression,
-                scroll = settings.scroll,
+                scroll = if (layout == EpubLayout.REFLOWABLE) {
+                    settings.scroll
+                } else {
+                    false
+                },
                 axis = if (settings.scroll && !settings.verticalText) {
                     Axis.VERTICAL
                 } else {