+++ readium/navigator/src/main/java/org/readium/r2/navigator/pager/R2EpubPageFragment.kt
@@ -159,6 +159,8 @@ internal class R2EpubPageFragment : Fragment() {
             }
         }
 
+        webView.disablePageTurnsWhileScrolling =
+            navigator?.config?.disablePageTurnsWhileScrolling ?: false
         webView.settings.javaScriptEnabled = true
         webView.isVerticalScrollBarEnabled = false
         webView.isHorizontalScrollBarEnabled = false