+++ readium/navigators/media/tts/src/main/java/org/readium/navigator/media/tts/session/TtsSessionAdapter.kt
@@ -85,7 +85,7 @@ internal class TtsSessionAdapter<E : TtsEngine.Error>(
     )
 
     init {
-        val streamType = Util.getStreamTypeForAudioUsage(audioAttributes.usage)
+        val streamType = audioAttributes.streamType
         streamVolumeManager.setStreamType(streamType)
     }
 
@@ -343,23 +343,8 @@ internal class TtsSessionAdapter<E : TtsEngine.Error>(
         ttsPlayer.nextUtterance()
     }
 
-    @Deprecated("Deprecated in Java", ReplaceWith("hasPreviousMediaItem()"))
-    override fun hasPrevious(): Boolean {
-        return hasPreviousMediaItem()
-    }
-
-    @Deprecated("Deprecated in Java", ReplaceWith("hasPreviousMediaItem()"))
-    override fun hasPreviousWindow(): Boolean {
-        return hasPreviousMediaItem()
-    }
-
     override fun hasPreviousMediaItem(): Boolean {
         return previousMediaItemIndex != INDEX_UNSET
-    }
-
-    @Deprecated("Deprecated in Java", ReplaceWith("TODO(\"Not yet implemented\")"))
-    override fun previous() {
-        seekToPreviousMediaItem()
     }
 
     @Deprecated("Deprecated in Java", ReplaceWith("TODO(\"Not yet implemented\")"))
@@ -467,7 +452,7 @@ internal class TtsSessionAdapter<E : TtsEngine.Error>(
     }
 
     override fun getTrackSelectionParameters(): TrackSelectionParameters {
-        return TrackSelectionParameters.Builder(application)
+        return TrackSelectionParameters.Builder()
             .build()
     }